refactor: change subscriptions and upgrade page
