+++ src/app/(dashboard)/orgs/join/page.tsx
@@ -12,7 +12,7 @@ export default async function Page() {
   const session = await getServerSession(authOptions);
 
   if (!session) {
-    redirect(
+    return redirect(
       Routes.login({
         redirect: "/orgs/join",
       }),
